+++ HISTORY.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Export `SupportedLanguages`
+- Export `WidgetId`, `ElementId` and `Container`
 
 ## [1.1.0] - 2023-02-06